Renamed functions and classes to differentiate between lines and shapes.  Added tests to ensure 0 radius circles function appropriately.
