+++ TheSadRogue.Primitives.UnitTests/LinesTests.cs
@@ -45,49 +45,49 @@ namespace SadRogue.Primitives.UnitTests
         };
 
         // Algorithms which are guaranteed to return items in order from start to finish.
-        private static readonly ShapeAlgorithms.LineAlgorithm[] s_orderedAlgorithms =
+        private static readonly Lines.Algorithm[] s_orderedAlgorithms =
         {
-            ShapeAlgorithms.LineAlgorithm.Bresenham,
-            ShapeAlgorithms.LineAlgorithm.DDA,
-            ShapeAlgorithms.LineAlgorithm.Orthogonal
+            Lines.Algorithm.Bresenham,
+            Lines.Algorithm.DDA,
+            Lines.Algorithm.Orthogonal
         };
 
         // Each line algorithm paired with how it defines adjacency/distance between points.
-        private static readonly (ShapeAlgorithms.LineAlgorithm, Distance distanceRule)[] s_adjacency =
+        private static readonly (Lines.Algorithm, Distance distanceRule)[] s_adjacency =
         {
-            (ShapeAlgorithms.LineAlgorithm.Bresenham, Distance.Chebyshev),
-            (ShapeAlgorithms.LineAlgorithm.DDA, Distance.Chebyshev),
-            (ShapeAlgorithms.LineAlgorithm.Orthogonal, Distance.Manhattan)
+            (Lines.Algorithm.Bresenham, Distance.Chebyshev),
+            (Lines.Algorithm.DDA, Distance.Chebyshev),
+            (Lines.Algorithm.Orthogonal, Distance.Manhattan)
         };
 
-        private static readonly ShapeAlgorithms.LineAlgorithm[] s_allLineAlgorithms = Enum.GetValues<ShapeAlgorithms.LineAlgorithm>().ToArray();
+        private static readonly Lines.Algorithm[] s_allLineAlgorithms = Enum.GetValues<Lines.Algorithm>().ToArray();
 
-        public static IEnumerable<(ShapeAlgorithms.LineAlgorithm algo, (Point start, Point end) points)> OrderedTestCases =
+        public static IEnumerable<(Lines.Algorithm algo, (Point start, Point end) points)> OrderedTestCases =
             s_orderedAlgorithms.Combinate(s_testLines);
 
-        public static IEnumerable<(ShapeAlgorithms.LineAlgorithm algo, Distance distanceRule, (Point start, Point end) points)>
+        public static IEnumerable<(Lines.Algorithm algo, Distance distanceRule, (Point start, Point end) points)>
             AllTestCasesWithDistance =
                 s_adjacency.Combinate(s_testLines);
 
-        public static IEnumerable<(ShapeAlgorithms.LineAlgorithm algo, (Point start, Point end) points)> AllTestCases =
+        public static IEnumerable<(Lines.Algorithm algo, (Point start, Point end) points)> AllTestCases =
             s_allLineAlgorithms.Combinate(s_testLines);
 
         #endregion
 
         [Theory]
         [MemberDataTuple(nameof(OrderedTestCases))]
-        public void LineOrderingTests(ShapeAlgorithms.LineAlgorithm algo, (Point start, Point end) points)
+        public void LineOrderingTests(Lines.Algorithm algo, (Point start, Point end) points)
         {
-            var line = ShapeAlgorithms.GetLine(points.start, points.end, algo).ToArray();
+            var line = Lines.GetLine(points.start, points.end, algo).ToArray();
             Assert.Equal(points.start, line[0]);
             Assert.Equal(points.end, line[^1]);
         }
 
         [Theory]
         [MemberDataTuple(nameof(AllTestCasesWithDistance))]
-        public void LineAdjacencyTests(ShapeAlgorithms.LineAlgorithm algo, Distance distanceRule, (Point start, Point end) points)
+        public void LineAdjacencyTests(Lines.Algorithm algo, Distance distanceRule, (Point start, Point end) points)
         {
-            var line = ShapeAlgorithms.GetLine(points.start, points.end, algo).ToArray();
+            var line = Lines.GetLine(points.start, points.end, algo).ToArray();
 
             for (int i = 1; i < line.Length; i++)
                 Assert.Equal(1, distanceRule.Calculate(line[i - 1], line[i]));
@@ -95,13 +95,13 @@ namespace SadRogue.Primitives.UnitTests
 
         [Theory]
         [MemberDataTuple(nameof(AllTestCases))]
-        public void LineBoundsTests(ShapeAlgorithms.LineAlgorithm algo, (Point start, Point end) points)
+        public void LineBoundsTests(Lines.Algorithm algo, (Point start, Point end) points)
         {
             var min = new Point(Math.Min(points.start.X, points.end.X), Math.Min(points.start.Y, points.end.Y));
             var max = new Point(Math.Max(points.start.X, points.end.X), Math.Max(points.start.Y, points.end.Y));
             var expectedBounds = new Rectangle(min, max);
 
-            var line = ShapeAlgorithms.GetLine(points.start, points.end, algo).ToArray();
+            var line = Lines.GetLine(points.start, points.end, algo).ToArray();
             foreach (var point in line)
                 Assert.True(expectedBounds.Contains(point));
         }
@@ -109,7 +109,7 @@ namespace SadRogue.Primitives.UnitTests
         [Fact]
         public void BadAlgorithmTest()
         {
-            Assert.Throws<ArgumentException>(() => ShapeAlgorithms.GetLine((1, 2), (3, 4), (ShapeAlgorithms.LineAlgorithm)100));
+            Assert.Throws<ArgumentException>(() => Lines.GetLine((1, 2), (3, 4), (Lines.Algorithm)100));
         }
     }
 }
